Fix warnings: Use intptr_t for property client data indices
Change integer types used for wxProperty client data to intptr_t for portability and correctness. Replaced int n with intptr_t n and reinterpret_cast<unsigned long> with reinterpret_cast<intptr_t> in NormalsGenDialog.cpp to avoid truncation or undefined behavior when converting between pointers and integers on 64-bit platforms.

+++ src/program/NormalsGenDialog.cpp
@@ -120,7 +120,7 @@ void NormalsGenDialog::doMoveUpLayer(wxCommandEvent& WXUNUSED(event)) {
 	if (!s || !s->IsCategory() || s->GetName() == "Background")
 		return;
 
-	int n = 1;
+	intptr_t n = 1;
 	for (auto it = pgLayers->GetIterator(wxPG_ITERATE_CATEGORIES); !it.AtEnd(); ++it) {
 		wxPGProperty* p = it.GetProperty();
 		p->SetClientData(reinterpret_cast<void*>(n++));
@@ -135,7 +135,7 @@ void NormalsGenDialog::doMoveUpLayer(wxCommandEvent& WXUNUSED(event)) {
 	if (!me || !prev)
 		return;
 
-	int t = reinterpret_cast<unsigned long>(me->GetClientData());
+	intptr_t t = reinterpret_cast<intptr_t>(me->GetClientData());
 	me->SetClientData(prev->GetClientData());
 	prev->SetClientData(reinterpret_cast<void*>(t));
